Enum abstract_ns::Error

pub enum Error {
    InvalidName(String, &'static str),
    TemporaryError(Box<StdError + Send + Sync>),
    NameNotFound,
}

A generic name resolution error

It's designed to provide basic abstraction over error types and also provide as much information as possible by carrying original error

Variants
InvalidName(String, &'static str)

Couldn't parse a name before resolution

It's expected that this error is permanent and is a failure of validating user input or the name in the configuration is invalid, but it's possible that some resolver have very specific requirements for names, so you might want to change resolver too.

TemporaryError(Box<StdError + Send + Sync>)

Temporary name resolution error

This means either name server returned this kind of error or we couldn't connect to a name server itself. It's safe to assume that you can retry name resolution in a moment

NameNotFound

We have sucessfully done name resolution but there is no such name

Methods

impl Error

fn into_io(self) -> IoError

Wraps the error into std::io::Error.